Duplicate accounts are no longer possible! :)
Duplicate accounts are now impossible, as I added a check which now produces an error if the user tries to make a duplicate account. It uses the stream.anyMatch() method, which loops through the arrayList, and searches for duplicates. If duplicates are found, it returns false, and the debug.error appears.